Update core dump processing for IDF v5

+++ src/freertos_drivers/esp32/Esp32CoreDumpUtil.hxx
@@ -88,22 +88,23 @@ public:
       {
         // Convert the core dump to a text file
         string core_dump_summary =
-            StringPrintf("Task:%s (%d) crashed at PC %08x\n",
-                         details.exc_task, details.exc_tcb, details.exc_pc);
+            StringPrintf("Task:%s (%" PRIu32 ") crashed at PC %08x\n",
+                         details.exc_task, details.exc_tcb,
+                         (int)details.exc_pc);
         core_dump_summary += StringPrintf("Registers:\n");
         for (size_t idx = 0; idx < 16; idx += 4)
         {
           core_dump_summary +=
               StringPrintf(
                   "A%02zu: 0x%08x A%02zu: 0x%08x A%02zu: 0x%08x A%02zu: 0x%08x\n",
-                  idx, details.ex_info.exc_a[idx],
-                  idx + 1, details.ex_info.exc_a[idx + 1],
-                  idx + 2, details.ex_info.exc_a[idx + 2],
-                  idx + 3, details.ex_info.exc_a[idx + 3]);
+                  idx, (int)details.ex_info.exc_a[idx],
+                  idx + 1, (int)details.ex_info.exc_a[idx + 1],
+                  idx + 2, (int)details.ex_info.exc_a[idx + 2],
+                  idx + 3, (int)details.ex_info.exc_a[idx + 3]);
         }
         core_dump_summary +=
             StringPrintf("EXCCAUSE: %08x EXCVADDR: %08x\n",
-                         details.ex_info.exc_cause, details.ex_info.exc_vaddr);
+                         (int)details.ex_info.exc_cause, (int)details.ex_info.exc_vaddr);
         if (details.ex_info.epcx_reg_bits)
         {
           core_dump_summary += "EPCX:";
@@ -112,7 +113,7 @@ public:
             if (details.ex_info.epcx_reg_bits & BIT(idx))
             {
               core_dump_summary +=
-                  StringPrintf("%zu:%08x ", idx, details.ex_info.epcx[idx]);
+                  StringPrintf("%zu:%08x ", idx, (int)details.ex_info.epcx[idx]);
             }
           }
           core_dump_summary += "\n";
@@ -121,7 +122,7 @@ public:
         for (size_t idx = 0; idx < details.exc_bt_info.depth; idx++)
         {
           core_dump_summary +=
-              StringPrintf(" 0x%08x", details.exc_bt_info.bt[idx]);
+              StringPrintf(" 0x%08x", (int)details.exc_bt_info.bt[idx]);
           if (details.exc_bt_info.corrupted)
           {
             core_dump_summary += "(corrupted)";
